Rename file with analysis function for consistency reasons
In this way we have freesound_audio_processing.py and freesound_audio_analysis.py as two complementary files
